--- client-src/overlay/runtime-error.js
@@ -1,50 +1,0 @@
-/**
- *
- * @param {Error} error
- */
-function parseErrorToStacks(error) {
-  if (!error || !(error instanceof Error)) {
-    throw new Error(`parseErrorToStacks expects Error object`);
-  }
-  if (typeof error.stack === "string") {
-    return error.stack
-      .split("\n")
-      .filter((stack) => stack !== `Error: ${error.message}`);
-  }
-}
-
-/**
- * @callback ErrorCallback
- * @param {ErrorEvent} error
- * @returns {void}
- */
-
-/**
- * @param {ErrorCallback} callback
- */
-function listenToRuntimeError(callback) {
-  window.addEventListener("error", callback);
-
-  return function cleanup() {
-    window.removeEventListener("error", callback);
-  };
-}
-
-/**
- * @callback UnhandledRejectionCallback
- * @param {PromiseRejectionEvent} rejectionEvent
- * @returns {void}
- */
-
-/**
- * @param {UnhandledRejectionCallback} callback
- */
-function listenToUnhandledRejection(callback) {
-  window.addEventListener("unhandledrejection", callback);
-
-  return function cleanup() {
-    window.removeEventListener("unhandledrejection", callback);
-  };
-}
-
-export { listenToRuntimeError, listenToUnhandledRejection, parseErrorToStacks };
