+++ package.json
@@ -61,7 +61,6 @@
     "connect-history-api-fallback": "^2.0.0",
     "express": "^4.21.1",
     "graceful-fs": "^4.2.6",
-    "html-entities": "^2.4.0",
     "http-proxy-middleware": "^2.0.7",
     "ipaddr.js": "^2.1.0",
     "launch-editor": "^2.6.1",
